Extra flag added to idmap cache API.
The uidmap_get() and other cache read functions can return ID_MAPPER_CACHE_EXPIRE.
This will happen in case, the cache entry is expired after timeout.
We want to reuse the memory structure of this entry once it is refreshed.
The entry is never removed from hash table. It is marked as expired.
When the overwrite flag is used, we check if the entry is present in hash table and reuse the memory.
This avoids a bunch of malloc/free at the expiry timeout.

Signed-off-by: suhrud <supatank@in.ibm.com>